Adds basic filerepository.
Gets tricky when filerepository has to change the db schema.
"save" on repo closes existing db connections, & creates a new one
with a new store corresponding to the file added.
This will be an issue with other open stores who share a connection
with the filerepo : it will have been closed. This has to be taken care
of.